Run bash script in createCommitObjects

+++ server/src/main/java/edu/purdue/cs/encourse/model/BarValue.java
@@ -14,8 +14,8 @@ import javax.persistence.Embeddable;
 @AllArgsConstructor
 public class BarValue {
 	
-	public byte index;
+	public int index;
 	
-	public short size;
+	public int size;
 	
 }
